Include the version number in the binaries
Created a small make script (make.go) that embeds
the latest tag, the date and the current git
hash in as the version number of the application.

The build no longer used gox, but can now be done
with the `make.go` script:

Compile myip:

    go run make.go -install:

Cross-Compile myip:

    go run make.go -crosscompile

Get help for the make script:

    go run make.go

The cross-compilation no longer supports x86 processor
architectures.

The cross-compilation process now creates specialized
binaries for ARM 5, 6 and 7.